- rename backend - remove basicConfig setting
