+++ src/visions/backends/pandas_/types/complex.py
@@ -1,9 +1,9 @@
 import pandas as pd
 from pandas.api import types as pdt
 
-from visions.backends.pandas_be import test_utils
-from visions.backends.pandas_be.series_utils import series_not_empty, series_not_sparse
-from visions.backends.pandas_be.types.float import string_is_float
+from visions.backends.pandas_ import test_utils
+from visions.backends.pandas_.series_utils import series_not_empty, series_not_sparse
+from visions.backends.pandas_.types.float import string_is_float
 from visions.types.complex import Complex
 from visions.types.string import String
