+++ src/visions/backends/pandas_/types/date_time.py
@@ -3,8 +3,8 @@ from functools import partial
 import pandas as pd
 from pandas.api import types as pdt
 
-from visions.backends.pandas_be import test_utils
-from visions.backends.pandas_be.series_utils import (
+from visions.backends.pandas_ import test_utils
+from visions.backends.pandas_.series_utils import (
     series_handle_nulls,
     series_not_empty,
     series_not_sparse,
