+++ src/visions/backends/pandas_/types/geometry.py
@@ -3,7 +3,7 @@ import sys
 
 import pandas as pd
 
-from visions.backends.pandas_be.series_utils import (
+from visions.backends.pandas_.series_utils import (
     series_handle_nulls,
     series_not_empty,
 )
