+++ src/visions/backends/pandas_/types/image.py
@@ -3,7 +3,7 @@ from pathlib import Path
 
 import pandas as pd
 
-from visions.backends.pandas_be.series_utils import (
+from visions.backends.pandas_.series_utils import (
     series_handle_nulls,
     series_not_empty,
 )
