+++ src/visions/backends/pandas_/types/integer.py
@@ -2,7 +2,7 @@ import numpy as np
 import pandas as pd
 from pandas.api import types as pdt
 
-from visions.backends.pandas_be.series_utils import (
+from visions.backends.pandas_.series_utils import (
     series_handle_nulls,
     series_not_empty,
     series_not_sparse,
