+++ src/visions/backends/pandas_/types/ip_address.py
@@ -2,8 +2,8 @@ from ipaddress import _BaseAddress, ip_address
 
 import pandas as pd
 
-from visions.backends.pandas_be import test_utils
-from visions.backends.pandas_be.series_utils import (
+from visions.backends.pandas_ import test_utils
+from visions.backends.pandas_.series_utils import (
     series_handle_nulls,
     series_not_empty,
 )
