+++ src/visions/backends/pandas_/types/path.py
@@ -2,7 +2,7 @@ import pathlib
 
 import pandas as pd
 
-from visions.backends.pandas_be.series_utils import (
+from visions.backends.pandas_.series_utils import (
     series_handle_nulls,
     series_not_empty,
 )
