+++ src/visions/backends/pandas_/types/time.py
@@ -2,7 +2,7 @@ from datetime import time
 
 import pandas as pd
 
-from visions.backends.pandas_be.series_utils import (
+from visions.backends.pandas_.series_utils import (
     class_name_attrs,
     series_handle_nulls,
     series_not_empty,
