+++ src/visions/backends/pandas_/types/url.py
@@ -2,7 +2,7 @@ from urllib.parse import ParseResult, urlparse
 
 import pandas as pd
 
-from visions.backends.pandas_be.series_utils import (
+from visions.backends.pandas_.series_utils import (
     isinstance_attrs,
     series_handle_nulls,
     series_not_empty,
