+++ src/visions/backends/pandas_/types/uuid.py
@@ -2,12 +2,12 @@ import uuid
 
 import pandas as pd
 
-from visions.backends.pandas_be.series_utils import (
+from visions.backends.pandas_.series_utils import (
     isinstance_attrs,
     series_handle_nulls,
     series_not_empty,
 )
-from visions.backends.pandas_be.test_utils import coercion_true_test
+from visions.backends.pandas_.test_utils import coercion_true_test
 from visions.types.string import String
 from visions.types.uuid import UUID
