--- src/visions/backends/pandas_be/__init__.py
@@ -1,1 +1,0 @@
-import visions.backends.pandas_be.types
