+++ src/visions/test/series.py
@@ -5,7 +5,7 @@ import pandas as pd
 
 def get_series() -> Dict[str, pd.Series]:
     from visions.backends.numpy_.sequences import get_sequences as get_numpy_sequences
-    from visions.backends.pandas_be.sequences import (
+    from visions.backends.pandas_.sequences import (
         get_sequences as get_pandas_sequences,
     )
     from visions.backends.python_.sequences import (
